+++ src/components/AlertBanner/__stories__/alertBanner.stories.helpers.js
@@ -1,10 +1,10 @@
-import { Link, Tip } from "vibe-storybook-components";
+import { StorybookLink, Tip } from "vibe-storybook-components";
 
 export const TipWhenToUse = () => (
   <Tip title="When to use?">
-    Alert banners should be reserved only for high-signal, system-level alert messages. For in-app notifications use a
-    <Link href="/?path=/docs/feedback-toast--docs" size={Link.sizes.SMALL}>
+    Alert banners should be reserved only for high-signal, system-level alert messages. For in-app notifications use a{" "}
+    <StorybookLink page="Feedback/Toast" size={StorybookLink.sizes.SMALL}>
       Toast.
-    </Link>
+    </StorybookLink>
   </Tip>
 );