+++ src/components/Skeleton/__stories__/skeleton.stories.helpers.js
@@ -1,10 +1,10 @@
-import { Link, Tip } from "vibe-storybook-components";
+import { StorybookLink, Tip } from "vibe-storybook-components";
 
 export const TipLoader = () => (
   <Tip>
-    Using loader after the pages are fully loaded consider using the
-    <Link href="/?path=/docs/feedback-loader--docs" size={Link.sizes.SMALL}>
+    Using loader after the pages are fully loaded consider using the{" "}
+    <StorybookLink page="Feedback/Loader" size={StorybookLink.sizes.SMALL}>
       loader component.
-    </Link>
+    </StorybookLink>
   </Tip>
 );