+++ src/hooks/useClickableProps/__stories__/useClickableProps.stories.helpers.js
@@ -1,11 +1,11 @@
-import { Link, Tip } from "vibe-storybook-components";
+import { StorybookLink, Tip } from "vibe-storybook-components";
 
 export const TipClickable = () => (
   <Tip title="Check out our component solution for this use case">
-    For more simple use cases, you also can use our
-    <Link size={Link.sizes.SMALL} href="/?path=/docs/accessibility-clickable--docs">
+    For more simple use cases, you also can use our{" "}
+    <StorybookLink size={StorybookLink.sizes.SMALL} page="Accessibility/Clickable">
       Clickable
-    </Link>
+    </StorybookLink>{" "}
     component wrapper.
   </Tip>
 );