+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## 2.69.1 (2023-12-06)
+
+#### Bug Fixes
+* [#1802](https://github.com/mondaycom/monday-ui-react-core/pull/1802) fix: changed toggle typography ([@shaharzil](https://github.com/shaharzil))
+
+## 2.69.0 (2023-12-06)
+
+#### New Features
+* [#1801](https://github.com/mondaycom/monday-ui-react-core/pull/1801) feat(EditableTypography): add onClick event ([@YossiSaadi](https://github.com/YossiSaadi))
+
+#### Documentation
+* [#1800](https://github.com/mondaycom/monday-ui-react-core/pull/1800) Internal links fixes ([@talkor](https://github.com/talkor))
+* [#1797](https://github.com/mondaycom/monday-ui-react-core/pull/1797) docs: replace internal links with StorybookLink ([@talkor](https://github.com/talkor))
+
+## 2.68.0 (2023-12-05)
+
+#### New Features
+* [#1798](https://github.com/mondaycom/monday-ui-react-core/pull/1798) feat(Chips): allow label to accept custom Element ([@YossiSaadi](https://github.com/YossiSaadi))
+* [#1790](https://github.com/mondaycom/monday-ui-react-core/pull/1790) feat(Toast): add loading prop  ([@YossiSaadi](https://github.com/YossiSaadi))
+
+#### Documentation
+* [#1796](https://github.com/mondaycom/monday-ui-react-core/pull/1796) docs(Tipseen): fix floating variation story ([@YossiSaadi](https://github.com/YossiSaadi))
+* [#1795](https://github.com/mondaycom/monday-ui-react-core/pull/1795) fix: re-enable sidebar tags ([@talkor](https://github.com/talkor))
+* [#1763](https://github.com/mondaycom/monday-ui-react-core/pull/1763) docs: Storybook 7 migrate ([@SergeyRoyt](https://github.com/SergeyRoyt))
+
 ## 2.67.0 (2023-11-30)
 
 #### New Features